Split device libraries apart
